added hashes

+++ filesort/collect/Collect.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"crypto/sha256"
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -34,26 +35,105 @@ type Record struct {
 }
 
 type Collect struct {
-	TxCount       int     // Transactions written
-	GuessCnt      int     // Just Info.  Count of guesses made to find hash
-	OffsetToIndex [8]byte // Offset to the Index Table
-	Offset        int64   // Keep the offset to Index Table as a uint64 as well
-	OffsetEnd     int64   // Offset end of file
-	Filename      string  // Name of the output file
-	TmpDirName    string  // Name of a temporary directory where we put the tmp files
+	TxCount       int               // Transactions written
+	GuessCnt      int               // Just Info.  Count of guesses made to find hash
+	OffsetToIndex [8]byte           // Offset to the Index Table
+	Offset        int64             // Keep the offset to Index Table as a uint64 as well
+	OffsetEnd     int64             // Offset end of file
+	Filename      string            // Name of the output file
+	TmpDirName    string            // Name of a temporary directory where we put the tmp files
+	outHash       *os.File          // File holding sorted, de-duplicated hashes
+	out           *os.File          // Place to collect, sort, and remove duplicates
+	tmpHashFiles  map[byte]*os.File // File pointing to the output file
+	tmpFiles      map[byte]*os.File // File pointing to all the temp files
+}
 
-	out      *os.File          // File pointing to the output file
-	tmpFiles map[byte]*os.File // File pointing to all the temp files
+// WriteHash
+// Sorts Hashes into buckets.
+// Allows them to be sorted, deduplicated, then written all into the hash file
+func (c *Collect) WriteHash(hash []byte) (err error) {
+	defer func() {
+		if msg := recover(); msg != nil {
+			err = fmt.Errorf("%v", msg) // Return panic's error
+		}
+	}()
+	_, err = c.outHash.Seek(0, io.SeekEnd)       // Position to the end of the file (get that index)
+	E(err, "failed to sync to EOF in writeHash") //
+	_, err = c.tmpHashFiles[hash[0]].Write(hash) // Write the record to the bucket indicated by the transaction hash
+	return nil
+}
+
+// BuildHashFile
+// Collect all the hashes from the byte hash files and build one sorted Hash File
+// Leaves the file pointer at the start of Hash file for easy serial access
+func (c *Collect) BuildHashFile() (err error) {
+	defer func() {
+		if msg := recover(); msg != nil {
+			err = fmt.Errorf("%v", msg) // Return panic's error
+		}
+	}()
+
+	var buffer [32 * 10240]byte //
+	var hashes [][32]byte       //
+	for i := 0; i < 256; i++ {  //         Run through all file bins
+		hashes = hashes[:0]
+		_, err := c.tmpHashFiles[byte(i)].Seek(0, io.SeekStart) // Seek to start of tmp file
+		E(err, "failed set 0 on tmp hash file")                 //
+		for {                                                   //
+			len, err := c.tmpHashFiles[byte(i)].Read(buffer[:]) // Read the buffer
+			EB(len%32 != 0, "read encountered partial hash")    // Going to ignore nasty edge cases if the
+			if errors.Is(err, io.EOF) {                         // If err is end of file break
+				break //
+			} //
+			for i := 0; i < len/32; i++ { //                       Run through the buffer and break up the hashes
+				var h [32]byte             //                      Put each hash into a slice to sort
+				copy(h[:], buffer[i*32:])  //
+				hashes = append(hashes, h) //
+			} //
+		} //                                                       After all the records in a bucket are collected
+		sort.Slice(hashes, func(i, j int) bool { //                  Sort them
+			less := bytes.Compare(hashes[i][:], hashes[j][:]) //
+			return less == -1
+		}) //
+		var last [32]byte
+		for _, h := range hashes { //                                Then write them all out to the end of
+			if !bytes.Equal(last[:], h[:]) {
+				_, e1 := c.outHash.Write(h[:])                    // every record in the
+				E(e1, "failed write record hash to transactions") // index table is a hash
+			}
+			last = h
+		} //                                                        then read the transaction.
+	}
+	c.outHash.Seek(0,io.SeekStart)	// Leave the file position at start of file
+	return nil
+}
+
+// GetHash
+// Get the next hash out of the hash file. After all the hashes have been put into collect,
+// the user then does a BuildHashFile().  Afterwards, the caller can walk serially through
+// the hashes by making repeated calls to GetHash.  Returns nil on EOF
+func (c *Collect) GetHash() []byte {
+	var h [32]byte
+	l, err := c.outHash.Read(h[:])
+	if errors.Is(err,io.EOF) {
+		return nil
+	}
+	E(err,"error reading hash")
+	EB(l != 32, "return from GetHash not 32 bytes")
+	return h[:]	
 }
 
 // NewCollect
 // The outputName is the full path of the desired output file.  If it exists,
 // the outputName will be deleted.
 //
+// If build is true, then structures to build a snapshot are initialized.  If false,
+// then the snapshot is simply opened to allow queries
+//
 // A directory of the name outputName.tmp will be created. After successful processing
 // the directory will be deleted.  If one exists when NewCollect is called, that directory
 // will first be deleted.
-func NewCollect(outputName string) (collect *Collect, err error) {
+func NewCollect(outputName string, build bool) (collect *Collect, err error) {
 	c := new(Collect)
 	defer func() {
 		if msg := recover(); msg != nil {
@@ -64,21 +144,37 @@ func NewCollect(outputName string) (collect *Collect, err error) {
 	}()
 	c.Filename = outputName
 	c.TmpDirName = outputName + ".tmp"
+	thd := c.TmpDirName + "/hashes"
 
-	c.out, err = os.Create(c.Filename) // create transactions
-	E(err, "create transactions fail") // error stuff
-	c.setLength()                      // write the offset to the index (length of all tx)
+	if !build {
+		return c, c.Open()
+	}
 
-	os.RemoveAll(c.TmpDirName)                // Remove any lingering temp directory
-	err = os.Mkdir(c.TmpDirName, os.ModePerm) // Create a new one
-	E(err, "failed to create tmp directory")  //
+	c.out, err = os.Create(c.Filename)                       // create transactions
+	E(err, "create transactions fail")                       // error stuff
+	c.setLength()                                            // write the offset to the index (length of all tx)
+	os.RemoveAll(c.TmpDirName)                               // Remove any lingering temp directory
+	err = os.Mkdir(c.TmpDirName, os.ModePerm)                // Create a new one
+	E(err, "failed to create tmp directory")                 //
+	err = os.Mkdir(thd, os.ModePerm)                         // Create a hashes directory
+	E(err, "error creating hashes tmp directory")            //
+	c.outHash, err = os.Create(fmt.Sprintf("%s/h.tmp", thd)) // Create tmp hash directory
+	E(err, "failed to create tmp hash file")                 //
+
+	c.tmpHashFiles = make(map[byte]*os.File, 256) // Create map for all 256 tmp hash files
+	for i := 0; i < 256; i++ {                    // Create a bucket file for every byte value
+		file, err := os.Create(fmt.Sprintf("%s/hash%03d", thd, i)) // Create each bucket
+		E(err, "creating temp files")                              //
+		c.tmpHashFiles[byte(i)] = file                             // Put in map based on index
+	}
+
 	c.tmpFiles = make(map[byte]*os.File, 256) // Create a map for all 256 tmp files
-
-	for i := 0; i < 256; i++ { // Create a bucket file for every byte value
+	for i := 0; i < 256; i++ {                // Create a bucket file for every byte value
 		file, err := os.Create(fmt.Sprintf("%s/sort%03d", c.TmpDirName, i)) // Create each bucket
 		E(err, "creating temp files")                                       //
 		c.tmpFiles[byte(i)] = file                                          // Put in map based on index
 	}
+
 	return c, nil
 }
 
@@ -180,7 +276,7 @@ func (c *Collect) SortIndexes() (err error) {
 			if len%40 != 0 {                                // Going to ignore nasty edge cases if the
 				panic("read a partial record") //               file system won't give me full buffers
 			} //                                                but check if it doesn't
-			if len == 0 && err != nil { //	                   If len is 0 and err is end of file break
+			if errors.Is(err, io.EOF) { //	                   If len is 0 and err is end of file break
 				break //
 			} //
 			var r Record                  //                   Assume we can hold a full bin in memory
@@ -193,7 +289,7 @@ func (c *Collect) SortIndexes() (err error) {
 		} //                                                         After all the records in a bucket are collected
 		sort.Slice(rs, func(i, j int) bool { //                      Sort them
 			less := bytes.Compare(rs[i].hash[:], rs[j].hash[:]) //
-			return less == 0
+			return less == -1
 		}) //
 		for _, r := range rs { //                                    Then write them all out to the end of
 			_, e1 := c.out.Write(r.hash[:])                      // every record in the
@@ -258,11 +354,13 @@ func (c *Collect) Fetch(index interface{}) (transaction, hash []byte, err error)
 		lb := int64(0)                                            // lower and upper bounds of the search for a hash
 		ub := (c.OffsetEnd - c.Offset) / 40                       //   that matches t.  The table is sorted by hash
 		ratio := float64(binary.BigEndian.Uint32(t)) / 0xFFFFFFFF // Hashes are evenly spread, so estimate location
+		step := float64(1) / float64(ub)                          // "weight" of being off by 1 for this many entries
+		_ = step                                                  //
 		indexRange := float64(ub - lb)                            // Range of indexes where the value might be
 		guess := int64(indexRange * ratio)                        // file by using the first 4 bytes of the hash sought
 
 		for {
-			if guess < lb || guess >= ub { //                        If the guess is out of bounds, t wasn't found
+			if lb == ub { //                        If the guess is out of bounds, t wasn't found
 				return transaction, hash, fmt.Errorf("transaction %x... not found", t[:8])
 			}
 			c.GuessCnt++   // Just count guesses in batches of roughly 10
@@ -278,24 +376,25 @@ func (c *Collect) Fetch(index interface{}) (transaction, hash []byte, err error)
 			E(err, "failed to seek into the index table")    //
 			_, err = c.out.Read(buff[:(e-b)*40])             // Read the number of entries (max 10) at 40 bytes each
 			E(err, "failed to access the index table")       //
+			var hashes [][]byte
+			for i := int64(0); i < e-b; i++ {
+				hashes = append(hashes, buff[i*40:i*40+32])
+			}
 		search:
 			for i := int64(0); i < e-b; i++ { // Read all the entries we read
-				fmt.Printf("t=%x\nb=%x\n\n", t, buff[i*40:i*40+32])
-				buff2 := buff[i*40 : i*40+32]
-				switch bytes.Compare(t, buff2) { //
+				switch bytes.Compare(t, hashes[i]) { //
 				case -1: //                                     Is t < buff+i? Gotta keep looking
-					lb = b + i + 1 //                           The lower bound is at least as high as next element
-					guess = lb + 5 //							The next guess is at least 5 above that
+					ub = b + i    //                            table hash is greater than t, a HIGHER bound
+					guess = b - 5 //							The next guess is at least 5 above that
 					break search
-						
 				case 0: //                                      Is t == buff, we have found the answer!
 					tx, hash, err := c.readTx(int(b + i)) //    Read the tx, and return what it gives us!
 					E(err, "transaction not found")       //
 					return tx, hash, err                  //    Return the tx
 
-				case 1: //                                      is t > buff
-					ub = b + i        //                        The upper bound is at least here
-					guess = b + i - 5 //                        The next guess will be 5 entries before that
+				case 1: //                                      is t > hash
+					lb = b + i + 1 //                           Hash found here is lower than t a LOWER bound
+					guess = e + 5  //                           The next guess will be 5 entries before that
 				}
 			}
 		}
@@ -303,12 +402,12 @@ func (c *Collect) Fetch(index interface{}) (transaction, hash []byte, err error)
 	panic("type not supported by c.Fetch()")
 }
 
-func (c *Collect)TestIndex(){
-	c.out.Seek(c.Offset,io.SeekStart)
+func (c *Collect) TestIndex() {
+	c.out.Seek(c.Offset, io.SeekStart)
 	var record [40]byte
 	for {
 		_, err := c.out.Read(record[:])
-		fmt.Printf("%x %8d\n",record[:8],binary.BigEndian.Uint64(record[32:]))
+		fmt.Printf("%x %8d\n", record[:8], binary.BigEndian.Uint64(record[32:]))
 		if err != nil {
 			break
 		}
